server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
pydantic==2.7.4
pydantic-settings==2.10.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
slowapi==0.1.9
langchain==1.0.3
langchain-google-genai==3.0.0
langchain-openai==1.0.1
langchain-anthropic==1.1.0
langchain-community==0.4.1
python-multipart==0.0.6
requests==2.32.5
beautifulsoup4==4.12.3
google-cloud-secret-manager
google-cloud-logging==3.9.0
pytest
pytest-asyncio
mypy
ruff

# Web検索機能の依存関係
httpx==0.27.0

# RAG機能の依存関係
faiss-cpu==1.12.0  # NumPy 2.x対応版
pypdf==3.17.4

# Payment/Subscription機能の依存関係
google-api-python-client==2.108.0  # Google Play Developer API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# JWT認証の依存関係
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OAuth State管理の依存関係
redis==5.0.1